+++ themes/uhsg_theme/js/degree_programme_switcher.js
@@ -1,17 +1,31 @@
 (function ($) {
   'use strict';
-
   Drupal.behaviors.degreeProgrammeSwitcher = {
     attach: function(context, settings) {
-      $('.block-views-blockdegree-programmes-block-1').once().find('.degree-programme-switcher__container').addClass('visually-hidden');
-      $('.degree-programme-switcher__toggle', '.degree-programme-switcher').once().on('click', function() {
-        var toggleText = $(this).hasClass('collapsed') ? Drupal.t('Change') : Drupal.t('Close');
-        $(this).toggleClass('collapsed')
-          .text(toggleText)
-          .next('.degree-programme-switcher__container')
-          .toggleClass('visually-hidden');
+      var degreeProgrammeSwitcher = '.degree-programme-switcher',
+          header = $('.degree-programme-switcher__header', degreeProgrammeSwitcher),
+          container = $('.degree-programme-switcher__container', degreeProgrammeSwitcher),
+          toggle = $('.degree-programme-switcher__toggle', degreeProgrammeSwitcher),
+          searchInput = $('input[name="name"]', degreeProgrammeSwitcher),
+          toggleClass = 'collapsed',
+          toggleTextOpen = Drupal.t('Change'),
+          toggleTextClosed = Drupal.t('Close');
+
+      // toggle collapsed when clicking header
+      header.once().on('click', function() {
+        container.toggleClass(toggleClass);
+        toggle.text(container.hasClass(toggleClass) ? toggleTextOpen : toggleTextClosed);
+        searchInput.focus();
       });
+
+      // close when clicking outside
+      $(document).once().on('click', function(e) {
+        var clickedOutside = $(e.target).parents(degreeProgrammeSwitcher).length == 0;
+        if (container.hasClass(toggleClass) && clickedOutside) {
+          container.removeClass(toggleClass);
+        }
+      })
+
     }
   };
-
 }(jQuery));